Update ParsePage.cs
Prevent regex exception on null input. I got it when the model for select didn't have an "iname" field before listSelectOptions was overridden. The reason was not clear until I traced and debugged every call in ParsePage. I'd much easier if I see a select with empty names in the browser.